Added Edit category code
Program at this stage is around 90% complete.
New Categories can be created and any EXISTING data in the categories can be edited.
New items cannot at this stage be added into the program.
Existing items cannot be deleted empty records will be created.
No checking to see if number only values are only storing numbers and items cannot be reordered.
Minor CSS changes need to be made.
Plans to make the program not use template files stored on the hard drive but use the data embedded into the program.